repr: change ArrayDimension lower_bound to isize
Postgres supports representing array lower bounds as negative
values, e.g.

SELECT ('[-10:-9]={6,7}'::int[])[-9];
----
7

However, we couldn't support this because ArrayDimension's
lower_bound was usize. We can easily swap this out for an isize,
though, because we are guaranteed that all previous values of
lower_bound fit into an isize (they were all 1usize).

+++ src/sql/src/plan/query.rs
@@ -3561,7 +3561,7 @@ fn plan_subscript_array(
     ecx: &ExprContext,
     expr: HirScalarExpr,
     positions: &[SubscriptPosition<Aug>],
-    offset: usize,
+    offset: i64,
 ) -> Result<CoercibleScalarExpr, PlanError> {
     let mut exprs = Vec::with_capacity(positions.len() + 1);
     exprs.push(expr);
